Admin Edit Product Gallery

+++ app/Http/Livewire/Admin/AdminProductComponent.php
@@ -18,7 +18,9 @@ class AdminProductComponent extends Component
         if($product->images){
             $images=explode(',',$product->images);
             foreach($images as $image){
+              if($image){
                 unlink('assets/images/products'.'/'.$image);
+              }
             }
         }
         $product->delete();
